Removed unused imports and added code comments

+++ src/test/java/org/globant/automation/model/PetResponseDTO.java
@@ -1,10 +1,10 @@
 package org.globant.automation.model;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 
 @Data
@@ -12,7 +12,7 @@ import java.util.ArrayList;
 @AllArgsConstructor
 @NoArgsConstructor
 public class PetResponseDTO {
-    private long id;
+    private BigInteger id;
     private CategoryDTO category;
     private String name;
     private ArrayList<String> photoUrls;
